+++ demo/lit-app-v7/src/my-element.ts
@@ -12,8 +12,8 @@ export type Variant =
   | "text";
 
 type DataObject = {
-  test: string;
-  value: string;
+  test?: string;
+  value?: string;
 };
 
 /**
@@ -36,7 +36,7 @@ export class MyElement extends LitElement {
   @property({ reflect: true }) variant: Variant = "primary";
 
   /** The button's theme variant. */
-  @property({ attribute: false }) dataObject?: DataObject;
+  @property({ attribute: false }) dataObject?: DataObject = {};
 
   /**
    * @deprecated replaced by `docs-hint`
@@ -151,24 +151,24 @@ export class MyElement extends LitElement {
       color: #535bf2;
     }
 
-    button {
-      border-radius: 8px;
-      border: 1px solid transparent;
-      padding: 0.6em 1.2em;
-      font-size: 1em;
-      font-weight: 500;
-      font-family: inherit;
-      background-color: #1a1a1a;
-      cursor: pointer;
-      transition: border-color 0.25s;
-    }
-    button:hover {
-      border-color: #646cff;
-    }
-    button:focus,
-    button:focus-visible {
-      outline: 4px auto -webkit-focus-ring-color;
-    }
+    // button {
+    //   border-radius: 8px;
+    //   border: 1px solid transparent;
+    //   padding: 0.6em 1.2em;
+    //   font-size: 1em;
+    //   font-weight: 500;
+    //   font-family: inherit;
+    //   background-color: #1a1a1a;
+    //   cursor: pointer;
+    //   transition: border-color 0.25s;
+    // }
+    // button:hover {
+    //   border-color: #646cff;
+    // }
+    // button:focus,
+    // button:focus-visible {
+    //   outline: 4px auto -webkit-focus-ring-color;
+    // }
 
     @media (prefers-color-scheme: light) {
       a:hover {